Cache file and embed reading fix/improvements
The cache should improve speed without the bot relying on it. The bot will load the cache as a dictionary upon starting. In the case of things that arent cached it will make a request to obtain the information (so this means in worst case you will get old speeds). Currently the cache will not account for duplicate message contents and will store them on completely seperate lines, this is to account for edited messages in the future. Also fixed the issue of the bot not reading the message if there is an embed message such as a youtube video link.